reorganized testing to put fixtures into the test_basics.py file, easier to access from elsewhere
